feat: refactor chat message internal note component and add delete functionality
- Updated the import paths for internal note components to a new directory structure.
- Introduced a new `ChatMessageInternalNote` component with a modal for deleting internal notes.
- Refactored message text formatting logic into a utility function for better reusability.
- Enhanced the `MessageText` component to utilize the new formatting function.
