+++ src/components/chats/Message/MessageText.vue
@@ -15,6 +15,8 @@
 </template>
 
 <script>
+import { formatMessageText } from '@/utils/string';
+
 export default {
   name: 'ChatsMessageText',
 
@@ -30,61 +32,7 @@ export default {
   },
   computed: {
     formattedText() {
-      function treatTextUrl(text) {
-        const urlRegex = /(https?:\/\/[^\s]+)/g;
-        return text.replace(urlRegex, '<a href="$1" target="_blank">$1</a>');
-      }
-
-      function removeHtmlDangerousContent(text) {
-        return text.replace(
-          /<(\/)?([^> ]+)( [^>]+)?>/gi,
-          ($1, $2 = '', $3, $4 = '') => {
-            if (['b', 'i', 'u', 'ul', 'li', 'br', 'div'].includes($3)) {
-              const complements = [];
-
-              for (const i of $4.matchAll(
-                /((?<name1>[^ =]+)="(?<value1>[^"]*)"|(?<name2>[^ =]+)='(?<value2>[^"]*)')/g,
-              )) {
-                const name = i.groups.name1 || i.groups.name2;
-                const value = i.groups.value1 || i.groups.value2;
-
-                if (name === 'style') {
-                  const styles = [];
-
-                  for (const j of value.matchAll(
-                    /(?<propertyName>[^:]+):(?<propertyValue>[^;]+);?/g,
-                  )) {
-                    if (
-                      j.groups.propertyName.toLowerCase().trim() ===
-                      'text-align'
-                    ) {
-                      styles.push(
-                        `${j.groups.propertyName
-                          .toLowerCase()
-                          .trim()}: ${j.groups.propertyValue.trim()}`,
-                      );
-                    }
-                  }
-
-                  complements.push(`style="${styles.join('; ')};"`);
-                }
-              }
-
-              return `<${$2}${$3}${
-                complements.length ? ` ${complements.join(' ')}` : ''
-              }>`;
-            }
-
-            return '';
-          },
-        );
-      }
-
-      const formattedText = treatTextUrl(
-        removeHtmlDangerousContent(this.text).trim()?.replace(/\n/g, '<br/>'),
-      );
-
-      return typeof this.text === 'string' ? formattedText : '';
+      return formatMessageText(this.text);
     },
   },
 };
